requirement.txt
pandas
numpy
google-genai
google-cloud-texttospeech